expr: refactor

+++ samples/1.ts
@@ -5,7 +5,7 @@ function printf(msg: string, value: int): void;
 // @ts-ignore
 function main(): int {
 
-    let a: int = 1;
+    let a: bigint = 1;
     let b: boolean = false;
 
     printf("Value is %d", (a - b));
